+++ package.json
@@ -20,7 +20,8 @@
     "react": "^18.2.0",
     "react-dom": "^18.2.0",
     "react-scripts": "5.0.1",
-    "typescript": "^4.8.3"
+    "typescript": "^4.8.3",
+    "yarn": "^1.22.19"
   },
   "scripts": {
     "start": "react-scripts start",
